third_party/WebKit/LayoutTests/http/tests/devtools/sources/debugger-ui/continue-to-location-markers-expected.txt

CONSOLE MESSAGE: line 30: 10
CONSOLE MESSAGE: line 31: 10
CONSOLE MESSAGE: line 32: 10
CONSOLE MESSAGE: 10
CONSOLE MESSAGE: line 37: 10
CONSOLE MESSAGE: line 40: 10
CONSOLE MESSAGE: line 43: 10
CONSOLE MESSAGE: line 52: 10
CONSOLE MESSAGE: line 54: 10
CONSOLE MESSAGE: line 56: 10
CONSOLE MESSAGE: line 58: 10
Tests that continue to location markers are computed correctly.

Set timer for test function.
CodeMirror-persist-highlight source-frame-async-step-in @30:16 = '(xxx, yyy)'
CodeMirror-persist-highlight source-frame-async-step-in @31:21 = '(xxx, yyy)'
CodeMirror-persist-highlight source-frame-async-step-in @32:21 = '(xxx, /*zzz*/ yyy /* xyz    */)'
CodeMirror-persist-highlight source-frame-async-step-in @33:24 = 'bar2'
CodeMirror-persist-highlight source-frame-async-step-in @34:24 = 'console'
CodeMirror-persist-highlight source-frame-async-step-in @35:24 = 'console'
CodeMirror-persist-highlight source-frame-async-step-in @36:16 = 'function'
CodeMirror-persist-highlight source-frame-async-step-in @39:19 = 'async /* comment */  function'
CodeMirror-persist-highlight source-frame-async-step-in @42:19 = 'async function'
CodeMirror-persist-highlight source-frame-async-step-in @45:16 = 'bar2'
CodeMirror-persist-highlight source-frame-async-step-in @46:16 = '()'
CodeMirror-persist-highlight source-frame-async-step-in @47:16 = 'async ()'
CodeMirror-persist-highlight source-frame-async-step-in @48:16 = 'async (x, y)'
CodeMirror-persist-highlight source-frame-async-step-in @49:15 = 'bar1'
CodeMirror-persist-highlight source-frame-async-step-in @51:16 = '('
CodeMirror-persist-highlight source-frame-async-step-in @53:16 = '('
CodeMirror-persist-highlight source-frame-async-step-in @55:16 = 'async ('
CodeMirror-persist-highlight source-frame-async-step-in @61:16 = '(bar2)'
CodeMirror-persist-highlight source-frame-async-step-in @62:16 = 'Promise'
CodeMirror-persist-highlight source-frame-async-step-in @63:16 = 'Promise'
CodeMirror-persist-highlight source-frame-async-step-in @63:41 = 'bar2'
CodeMirror-persist-highlight source-frame-async-step-in @64:16 = '('
CodeMirror-persist-highlight source-frame-async-step-in @68:18 = 'bar2'
CodeMirror-persist-highlight source-frame-async-step-in @69:16 = 'bar2'
CodeMirror-persist-highlight source-frame-async-step-in @71:30 = 'bar2'
CodeMirror-persist-highlight source-frame-continue-to-location @24:12 = 'foo1'
CodeMirror-persist-highlight source-frame-continue-to-location @24:21 = 'foo1'
CodeMirror-persist-highlight source-frame-continue-to-location @25:12 = 'foo2'
CodeMirror-persist-highlight source-frame-continue-to-location @27:12 = 'a'
CodeMirror-persist-highlight source-frame-continue-to-location @30:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @30:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @31:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @31:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @32:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @32:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @33:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @33:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @34:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @34:32 = 'log'
CodeMirror-persist-highlight source-frame-continue-to-location @34:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @35:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @35:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @36:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @36:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @39:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @39:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @42:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @42:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @45:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @45:21 = 'bind'
CodeMirror-persist-highlight source-frame-continue-to-location @45:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @46:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @46:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @47:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @47:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @48:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @48:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @49:4 = 'setTimeout'
CodeMirror-persist-highlight source-frame-continue-to-location @50:14 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @51:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @51:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @53:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @53:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @55:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @55:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @57:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @57:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @59:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @59:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @61:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @61:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @62:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @62:24 = 'resolve'
CodeMirror-persist-highlight source-frame-continue-to-location @62:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @63:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @63:24 = 'resolve'
CodeMirror-persist-highlight source-frame-continue-to-location @63:36 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @63:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @64:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @64:25 = 'resolve'
CodeMirror-persist-highlight source-frame-continue-to-location @64:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @68:13 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @68:6 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @69:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @69:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @71:11 = 'then'
CodeMirror-persist-highlight source-frame-continue-to-location @71:35 = 'bind'
CodeMirror-persist-highlight source-frame-continue-to-location @71:4 = 'bar1'
CodeMirror-persist-highlight source-frame-continue-to-location @72:4 = 'return'
